Second navigation
Created a navigation bar.

Used flex with space-around to justify the label and the three dropdown menus.

Used the flex container "menu" to hold just the three menus.

Created a list of links for each one of the menus and set the display to none to hide it.

Set the lists to appear when the menus were hovered over.

Set the background to become light blue and the text to become bold on the button and the background to become light gray and the text to become bold in the list when they are hovered over.