Second button (pause)
Gave the button a thick double left border while removing all the other borders so all that is visible is the two pause lines.

Put these lines inside a div with a border and set the border-radius to 50% to make it a circle.

Set both the lines and the border to turn light blue when hovered over.