Third button
Created a label container to house this button.

Set the input to "checkbox."

Put a span inside the label.

Set the checkbox to display: none to hide it.

Gave the label and the span all the same dimensions.

Used :before to create a circle next to the span.

Set the circle to position: absolute and centered it inside the span.

Made the span blue and set it to turn light blue when the checkbox was checked.

Used translateX to make the circle toggle to the right when the checkbox was checked.

Set this transition to take .5 seconds.