Fourth button
Created a label container to hide this button.

Gave the button a box-shadow and used the shadow spread function to make it expand in all directions.

Gave the button another box-shadow that does not expand but extends to the right and bottom.

Set the dimensions of the label so it wrapped around the box-shadows in addition to the button itself.

Set the button to scale down to 95% and shift slightly down and to the right when clicked on.

Set the box-shadows to expand to twice their size and become half as opaque when the button is clicked on.