Second typography
Gave the background of the text a linear gradient of 240 degrees from light blue to blue.

Clipped the background to fit inside the text and made the text transparent so the gradient can be seen through it.

Wrapped a div around the text paragraph.

Gave the background of the div a linear gradient of 120 degrees from light blue to blue.

Put a very thin white drop-shadow around the text to make it easier to differentiate between the blues of the text and background toward the middle of the gradient.