First media card layout
Created a div for the media card and set its dimensions.

Put a h3 header inside the media card for the title.

Put a div inside the media card to represent a hypothetical image.

Put a paragraph inside the media card to hold the main text.

Put a label inside the media card that will house a button.