Second media card buttons
Moved the div containing the buttons to the bottom and set its dimensions to match the media card.

Set its background-color to blue.

Put the back button inside this div.

Created another button with a forward arrow instead of a back arrow.

Used relative positioning to move it to the right side.