Media query (550px-767px)
Gave the first navigation bar 3 columns instead of 6 to prevent the items form overflowing.

Set the second navigation bar to display: grid.

Put the menu label on the first row and spaced out the items on the second row of this grid.

Changed the flex-direction of the typography to column and centered the text.

Changed the flex-direction of the media cards to column and centered them.